Add mutation APIs for review groups and their users.
This adds APIs for creating, updating, and destroying review groups, and
for managing the user list on groups.

The /groups/ API responds to HTTP POST and DELETE. /groups/<name>/
responds to PUT for updating all fields.

The /groups/<name>/users/ resource responds to HTTP POST and accepts a
username to add to the group. /groups/<name>/users/<username>/ accepts a
HTTP DELETE to remove from the group.

Unit tests were added to ensure that all responses are as expected for
various conditions, and that they work properly with LocalSites.

This was originally based on work done by Stephen Holsappie and Paul Hobb.