Remove useless, broken supports_change_timezone conditionals.
I guess when we initially did the timezone support, we thought it might be
connected to the auth backend. The result of this was code that would hide the
field on the preferences dialog for all auth backends except the standard
builtin one, and then fail to save when users tried to change their preferences
because the field was still marked as required.

Because the timezone setting is just part of the user profile and unrelated to
the auth backend, there's no reason to have these conditionals present.

Fixes bug 2797.
Reviewed at http://reviews.reviewboard.org/r/3491/