Add API support for creating a review request with no repository.
Our API wasn't allowing the creation of review requests without a
repository. This should be allowed, as the web UI allows it, and not all
review requests going forward will have anything to do with diffs.

The 'repository' parameter is now optional. Documentation was updated to
explain the new usage of this parameter.