Change the FileAttachmentComment API to require an "extra_data." prefix for extra data.
Recently, FileAttachmentComment was given support for specifying extra
data to store with a comment. This took any unknown variable and stored
it as extra data, but that meant that typos would also get stored.

Now it requires an "extra_data." prefix for any variables meant to be
stored there. That corresponds to the resulting payload, which is a
dictionary in 'extra_data'.