Choose a better default for the hosting account on the repository form.
When switching hosting services on the repository form, the Account
dropdown would default to "<Link a new account>", unless there was
already a matching account selected before. This is not a good default
when there are actual, valid accounts previously linked.

Now we choose the 1st real account as the default, unless we find a
previously selected account in the list, in which case that'll be the
default.