Fix the style for the comment save bubble in the diff viewer.
A CSS rule was introduced for the floating collapse div in the diff
viewer that broke the style for the comment save bubble. The border of
the bubble was no longer covering all the text, which looked very
broken.

The div for the collapse button has been given a class and the offending
rule has been updated to specifically make use of that class.