Fix review request field sizing issues and hide unused fields.
Review request fields were never shown very well. They had a hacky setup
that caused them to jump when the content changed, such as more text or
an inline editor being shown. This was ugly and I just won't stand it
anymore.

Now instead of a 4 column approach (which was really the core issue, as
it meant that the table was reevaluating its layout based on any new
content all the time and causing things to jump), we have two tables
floated next to each other, each taking up 50% of the space.

These tables will no longer intrude on each others' space, preventing
things from jumping around.

A nice side-benefit is that when resized really small (such as on mobile
devices), the tables will themselves wrap, moving to a single column.

Along with this, I've changed the template to not show the Change Number
field unless one is present, and not show Repository unless it's a
repository-backed review request. This simplifies things in certain
cases.