Switch over to file attachments entirely in place of screenshot uploading.
This removes the ability to upload screenshots, and instead moves to
using file attachments for everything.

Dragging-and-dropping a file over the review request will just have a
big "Drop to Upload" link, instead of "Drop Screenshot" and
"Drop File Attachment."

The "Upload Screenshot" link on a review request is gone.

The entry for a non-repository review request now says
"(None - File attachments only)" instead of "(None - Graphics only)".