Fix the examples for resource payloads in the webapi docs.
We were showing examples for every allowed mimetype for a resource, but
this meant that we ended up showing both the generic (application/json)
and specific (application/vnd.reviewboard.org.<resource>+json)
mimetypes. To make it worse, we didn't apply proper formatting to the
latter.

We now filter out the generic mimetypes that we know we'll have
specifics for, and make sure to be smart about when to apply formatting
to mimetypes.