Immediately display the reply draft banner after adding a comment.
The reply draft banner wasn't first showing up until after scrolling or
resizing the window. This was due to the banner not having a container
until later in the call stack. To work around this, we just need to
defer the initial call to update the position until the call stack
unwinds, since we'll then have a container and can make the proper
computations.