Fix errors when trying to download raw diffs with unicode filenames.
The filename field in the header would cause errors if the diff that was
uploaded had characters that weren't US-ASCII. This change adds an encoding step
that strips out any offending characters.

While I was in here, I changed the silly "bug%s" logic for filename fallbacks to
use rb-<id>.patch.

Testing done:
Created a diff named test𝄡.diff and uploaded it. Downloaded the resulting raw
diff and saw it try to save with a filename of "test.diff"

Fixes bug 2581