Fix overzealous caching of static media.
When upgrading a Review Board site, the static media references would
often be pointing to the wrong places. This was because Django cached
the resulting media files in memcached, which is definitely not what we
want. However, it will first try a "staticfiles" backend, which we can
dynamically add and point to local memory instead.

We do this after we've loaded the rest of the settings, ensuring we
have settings.CACHES set up.