Add review-ui URL field to FileAttachmentResource.
In order to properly link to the review UIs when adding file attachments through
javascript, I need to have a URL. If the given attachment doesn't have a review
UI, this URL will just be an empty string.

Testing done:
Set a breakpoint in the javascript that runs after a file attachment succeeds
and looked at the resulting object.

Reviewed at http://reviews.reviewboard.org/r/3576/