Fix fallback mimetype handler.
One small issue crept in with some of the recent changes to add new handlers for
additional mimetypes: if a type didn't have a handler, it wouldn't fallback to
the generic one. This meant that uploading files without handlers would generate
errors trying to get things like the icon url or the screenshot.

Trivial fix.