Show draft captions in file attachment review UIs.
When an image (or other reviewable attachment) had a draft caption, it wouldn't
be shown in the review UI header. Fixing this was a pretty simple addition of
the same fallback logic that we use elsewhere.

Testing done:
Set a draft caption on an image attachment and loaded the review UI.

Reviewed at http://reviews.reviewboard.org/r/3606/