Stop using the blesscss service.
This was a cool thing while it lasted, but all of the node.js packages have
changed their APIs (because apparently javascript developers don't understand
what "API" means) and it's currently so broken that I just stopped the app on
cloudfoundry. Since we have support for local lessc now, and will be requiring
node.js for other javascript compilation steps to build the egg, we might as
well just depend on it.

Testing done:
python setup.py bdist_egg release