Prevent breakages when loading an old image file attachment comment.
ImageReviewUI introduced support for adding comments to a region of an
image file attachment, but any older comments loaded wouldn't have
region information. We assumed it was there, which caused us to break
with older comments. We now catch the error and just ignore that
particular comment when rendering the review UI.