Support the p4api 12.2.
This adds support for p4api 12.2. The main differences that affect us is
that P4Error was renamed to P4Exception, and there's a different error
when we can't reach a server (for repository not found).

This change adds compatibility for the new API, which we'll be able to
switch to later, hopefully successfully. It also keeps compatibility
with the old API for existing users.