Update our settings documentation.
This updates our settings documentation to cover all the changes that
were made in 1.7 (and even some that we made late into 1.6). That
includes new settings like the cache fields and max diff size.

The optimization page also got some changes. It was referencing
CACHE_BACKEND, which is gone, and so was updated to point to the Cache
Settings. While I was there, I added an important note to the web server
scale-out docs.