Add documentation for the user settings page.
This adds a page to the "Getting Started" section of the user manual explaining
how to get to the "My Account" page, and what the various settings mean. It also
explains what it means to join a group and how to set up your gravatar.

Testing done:
Build html docs (without styling) and looked at the output. Checked that links
worked.