Anonymous Access not synced with settings.
Even though the value for "anonymous read-only access" is stored in the
database properly, the display always shows it as unchecked.

This is because of the way SiteSettingsForm resets the value that is
setup by AuthenticationSettingsForm's load method. Calling the
super.load first followed by setting this attribute's initial value
fixes the problem.

Fixes bug #2787
Reviewed at http://reviews.reviewboard.org/r/3641/