Improve button styling in the admin UI.
This improves our button styling to be a bit consistent, nicer, and to
be closer to what we have on RBCommons.

As part of this, we're bringing back styling for the default/primary
buttons on admin UI forms.

Later, we may want to look into making this a standard styling across
the whole app.