Fiddle with pointer cursors in the diffviewer.
The diffviewer had some odd javascript that added a "cursor: pointer" style to
entire rows. This was a little silly and didn't apply universally. This change
removes that javascript, but adds in explicit css for keeping the hand pointer
cursor for the line numbers.

Testing done: Hovered all over areas in the diff viewer and checked the cursors.
Reviewed at http://reviews.reviewboard.org/r/3650/