Use the new button styles throughout the UI.
This introduces the new button styles everywhere, not just the admin UI.
The style is nice and we can guarantee consistency across
platforms/browsers.

The primary/danger styles aren't used much. I tried playing with them
for the different Publish/Save/Discard buttons, but all the colors ended
up looking a little distracting. So, we're just using the default style
in the majority of cases (the exception being the submit buttons on the
My Account page and New Review Request page).