Change the styling a bit for the per-comment issue displays.
This polishes up the per-comment issue status displays. The main change
is that the bordered "Issues" link with an icon has been changed to a
simple text link that blends in nicely with the color scheme of the
issue banner for the current state. This was an attempt to make it less
distracting. It's also been changed to "Show all issues".

The open issue yellow color is now a bit more faded, and less
in-your-face.

There's a very faint border on the resolved state, keeping a consistency
with the other states. It helps keep the content from jumping when
toggling the border.

The issue text is no longer shrunk. It's the standard text size. It just
felt small compared to the buttons.

There's also now more room before the first button, making it less
bunched together.