Improve the styling of file attachment thumbnails containing text.
File attachment thumbnails containing text had very small fonts, and in
some cases, had bordered content. The content was also all blue and
underlined, since it was in a link, and was not as wide as it should be.

This change shows small (but not super tiny) text, black, with a decent
width.