Add file attachment comments to e-mails.
Our e-mails never gained file attachment support. They weren't listed in
the review requests or the reviews or replies.

This adds that support, mostly. There's no thumbnailing support, because
we can't rely on the CSS to be correct or the URLs to be absolute. That
will require some new support that we'd have to think about.