Port the issue summary table to Backbone.js.
This takes the issue summary table code and moves it out of reviews.js
and into separate model and view files.

The logic is unchanged. The only changes are the move to Backbone files,
some renamed variables, and the use of Backbone events instead of using
jQuery.

Future changes will end up decoupling bits of these objects as we move to
Backbone representations for other parts of our codebase.