Add better explanations and output for rb-site and Manual Updates.
rb-site and the Manual Updates pages didn't go into the <Location> field
changes that were needed, which ended up confusing users. These are now
covered, as are equivalent lighttpd changes.

The rb-site resolve-check command was also a little confusing in that a
second resolve would say that there's no known check. Now it says that
the check was already resolved.

There's also a few style fixes so the manual update page is more
readable in general, and particularly when the stylesheet can't be
loaded.