Add JS utility functions for working with Backbone properties.
This introduces $.fn.bindProperty and $.fn.bindVisibility, which makes it
easy to bind DOM element properties with Backbone properties. They can
be set to update any element property or any model property, and the
direction of synchronization (model-to-element, element-to-model, or
both) can be customized. Values can also be inversed (useful when
setting "disable" on an element from a "can*" on a model).

This also introduces unit tests for all the options and scenarios.