Provide the foundation for JavaScript unit tests.
This introduces Jasmine for testing JavaScript code. Jasmine is a pretty
useful and popular JavaScript unit test library that can be used through
either a browser or node.js.

We supply a /js-tests/ URL when DEBUG is turned on, which will run the
test suite.

There are currently no tests, but we can start adding them by
introducing test files and bundling them into the js-tests JavaScript
bundle in settings.py. These should be in the rb/js/tests/ directory.