Add fallbacks for important console.* methods.
We can't always rely upon console.log and console.assert existing in the
browser. What's worse is that we can't rely on console.assert behaving
the same way. In Chrome, for instance, a failed assert will log but will
continue on in the function, causing further problems.

We now have an empty fallback for console.log, and a console.assert
wrapper that ensures we throw an exception on a failed assert.

This helps with consistency, and with our unit tests.