Add support for moved file handling for Perforce.
This adds support for representing moved files properly in Perforce.
This handles both files moved with and without changes made.

An accompanying change to post-review is needed in order to add moved
information to the diff format. Like the other fields in Perforce diffs
we respond to, this is entirely custom, as Perforce doesn't have any
useful information in p4 diff to provide.

This feature is funded by NetApp, Inc.