Convert the drag-and-drop support to Backbone.js views.
The file attachment DnD code has moved to DnDUploader and
DnDOverlayView objects.

DnDUploader handles listening for incoming dragenter events, creating the
overlay, and listening for file drops on the overlay.

DnDOverlayView displays the "Drop to Upload" overlay. It works just like
the previous one did, but the code has been made much simpler. We used
to have to deal with two drop boxes (screenshots and file attachments),
which complicated DnD events quite a bit (since we'd get dragleaves when
moving between elements). Now it's one element completely (a div with
text), so much of the heuristics have been able to be removed, including
the delay in fading out the overlay.

The CSS has been moved out to a dndUploader.less file, allowing us to
stay better organized.