Fix a small vulnerability in the file change list in the diff viewer.
A change from a while back for some reason accidentally introduced
making entries in the file change list "safe", meaning that any HTML in
the filename would be rendered in the page instead of escaped. This was
due to these using the escapejs filter.

I can't see any reason why we'd want to use this filter in this case, so
I'm removing it, fixing the vulnerability.