Properly normalize Perforce diffs with unchanged, moved files.
This was part of my original change, but got left out in the version
that was pushed upstream.

PerforceTool.normalize_patch looks to see if the diff being processed
represents an unmoved, changed file. The header we use for this isn't an
actual diff, and will make patch upset, so we normalize it to a blank
patch.