Fix file attachment names in changedescs.
When file attachments are added with no captions, the change descriptions would
just say "Added ". This is pretty horrible. This change adds a new property
which will prefer the caption and fall back on the basename.

Testing done:
Added a file without a filename and published. Checked that the change description
text was the file basename. Gave the file a caption, and then removed it.
Checked that the change description text was the caption.