Fix static media URLs when running the devel server with a SITE_ROOT.
The way that static media URLs ('/static/' and '/media/') were mapped when a
SITE_ROOT was defined was problematic, in that it was appending the SITE_ROOT
twice in the urlspec. The problem is that all the URLs were created with the
root prepended, but the way that settings.py defined the urls didn't have the
right distinctions between the names of the directories and the URL paths.

Testing done:
Tested that loading the dashboard included all static media with a SITE_ROOT of
both '/' and '/example/'.