Support localized revision strings in svn diffs.
If the host is running in a non-C locale, 'svn diff' will produce a diff with
the revision strings localized. Unsurprisingly, this would fail to parse
correctly. post-review already handles this by always invoking svn with LANG=C,
but if someone uploaded a diff manually, it would still break. This change adds
all the different localizations of "revision" to the regular expression that we
use for parsing.

Testing done: ran added unit tests

Reviewed at http://reviews.reviewboard.org/r/3749/