Add a capability to indicate support for Perforce moved files.
1.7.2 introduced moved file support for Perforce, but there wasn't a
good way for RBTools to really detect this that retained compatibility
with previous versions of Review Board. The reason is that we relied
upon the existing general diffs.moved_files capability, causing RBTools
to generate diffs for older releases that ended up being incompatible.

This introduces a scmtools.perforce.moved_files cap that can be used
instead. 1.7.3 will be the minimum version of Review Board that will
support moved files for Perforce with RBTools.