Fix setting time_added when publishing a review request.
The change added to set time_added to the publish time caused the
creation time to always be the publish time for every publish, not just
the first.

We now only set this when there hasn't been any changes previously made
to the review request and it's being published the first time.