Fix unescaping comment text in the diff viewer when the text is blank.
If the comment text was blank, we'd hit a failure when unescaping the
text. I believe this may have been a change in jquery when passing in an
empty string to html().

Now we only do the unescaping logic if the comment text is blank.

Ideally, this would use _.unescape, but we don't have a new enough
underscore.js yet.