Rewrite the CommentReply objects to be based on Backbone.js.
The new objects perform the same way the old ones did, but share much of
the code and make use of the BaseResource model.

There's a BaseCommentReply that does all the actual work, and subclasses
for diffs, file attachments, and screenshots. All the actual logic is
common, and the subclasses only provide the 'rspNamespace' key used to
parse the payload from the server.

Unit tests were added to cover the base serialization/deserialization
and validation.