Include repo name in review request emails
For Issue 2483. The review request email notification template was altered to
include the repository name along with the review group. This was requested
since it would be useful for groups with multiple repositories.

The template was altered so that if a repository is associated with the review
request, then the name will appear in parentheses after the review group. If no
repository was included, then nothing new will appear.

Testing done:
I set up emails to be sent through gmail smtp on my local test server and also
created a repository to test with. I then created two review requests, one with
the test repo and one with no repo.

The emails for the test repo review request sent emails saying 'Review Request
for Katkat (TestRepo) ' and the emails for the review request with no repo sent
emails saying 'Review Request for Katkat'. This was the expected result.

Fixes bug 2483.
Reviewed at http://reviews.reviewboard.org/r/3768/