Brand new user sees "None" for all dashboard counts in a local site.
Fix the Starred Public Review Count for new user

The CounterField initializer for starred_public_request_count was calculating
the count as 0 if the pk of that table was not null. The parantheses for this
calculation was not setup properly , which caused the value to be setup as None
for the first time user.

Also fixed a test which was assuming that the first time user will see None for
the counts. With this and the associated djblets fix, this isnt true anymore.
This fixes the reviewboard side of things and is associated to the review
request 3841 (which is a djblets fix).

Testing done: Manual testing with localsites and regular ones. Ran all
reviewboard tests.

Fixes bug 2873.
Reviewed at http://reviews.reviewboard.org/r/3842/