Put a clickable overlay on top of file attachment thumbnails.
File attachment thumbnails can theoretically include HTML with links (for
example, markdown or restructured text types). In these cases, wrapping the
thumbnail in a big <a> tag isn't sufficient to make sure everything links to the
review UI. This change puts a new block element on top of the thumbnail contents
for users to click on.

Testing done:
Uploaded a variety of file attachments, through both drag-and-drop and the
upload dialog. Checked that the overlay was clickable and covered any links
present in the thumbnail element.

Reviewed at http://reviews.reviewboard.org/r/3844/